// src/utils/enums.ts
export enum Classification {
    Mampu = "Mampu",
    Menengah = "Menengah",
    Kurang_Mampu = "Kurang_Mampu"
}

export enum HomeOwnership {
    Owned = "Milik Sendiri",
    Rented = "Sewa",
    Family = "Milik Keluarga"
}

export enum StockType {
    IN = "IN",
    OUT = "OUT"
} 

export enum TransactionStatus {
    Selesai = "Selesai",
    Menunggu = "Menunggu",
    Gagal = "Gagal"
}